Big Progress.
Decided to make the whole thing in a single main.java file instead of a
main and setup file. Added multiple methods within the program for
different aspects of the game itself such as: movement, checking and
performing unit upgrades, killing a unit and checking if said unit was
the last unit on that side of the board, ect. Also added more comments
overall.