First limits on moves
Made a method to calculate the possible moves of a specific piece as
well as a list of moves that a piece can do to kill a unit. Tried to
implement both into the program as well as implementing the rule that
"if you can take a piece, you must". Currently quite buggy so i need to
spend some time figuring out what went wrong/trying other ideas.